Successfully send requests changepw, getcert

+++ client/base64.h
@@ -5,7 +5,7 @@
 #include <string>
 typedef uint8_t BYTE;
 
-std::string base64_encode(BYTE const* buf, unsigned int bufLen);
+std::string base64_encode(BYTE const* buf, unsigned long bufLen);
 std::vector<BYTE> base64_decode(std::string const&);
 
 #endif
